% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{A barplot}
\usage{
barplot(
  data,
  x,
  colour_by = NULL,
  label_x = NULL,
  title = NULL,
  fill_col = "black",
  line_col = "black",
  horizontal = FALSE
)
}
\arguments{
\item{data}{The data to plot}

\item{x}{The column to use for the x axis}

\item{colour_by}{The column to colour the bars by (optional)}

\item{label_x}{A label for the x axis}

\item{title}{A title for the plot}

\item{fill_col}{The colour of the bars}

\item{line_col}{The colour of the outline of the bars}

\item{horizontal}{Whether the barplot should be horizontal (TRUE or FALSE)}
}
\value{
A plot
}
\description{
A barplot
}
\examples{
barplot(mpg, x = class)
barplot(mpg, x = manufacturer, colour_by = class)
barplot(mpg, x = manufacturer, colour_by = class, horizontal = TRUE, title = "Horizontal barplot")
}
