% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{A boxplot}
\usage{
boxplot(
  data,
  x,
  y = NULL,
  colour_by = NULL,
  label_x = NULL,
  label_y = NULL,
  title = NULL,
  fill_col = "white",
  line_col = "black",
  width = 0.7
)
}
\arguments{
\item{data}{The data to plot}

\item{x}{The column to use for the x axis}

\item{y}{The column to use for the y axis}

\item{colour_by}{The column to colour the boxes by (optional)}

\item{label_x}{A label for the x axis}

\item{label_y}{A label for the y axis}

\item{title}{A title for the plot}

\item{fill_col}{The colour of the box}

\item{line_col}{The colour of the outline of the box}

\item{width}{A number 0-1 representing the width of the boxplot}
}
\value{
A plot
}
\description{
A boxplot
}
\examples{
boxplot(mpg, x = displ)
boxplot(mpg, x = displ, y = class)
boxplot(mpg, x = displ, y = class, fill_col = "gold", title = "My boxplot")
}
