% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rladiesmelbITAS-package.R
\docType{package}
\name{rladiesmelbITAS-package}
\alias{rladiesmelbITAS}
\alias{rladiesmelbITAS-package}
\title{rladiesmelbITAS: Functions and data for RLadies Melbourne It Takes a Spark workshop}
\description{
A package to store helper functions and data for our annual event at It Takes A Spark - Become a Disease Detective - a workshop to introduce school students to coding, R and epidemiology. These helper functions are mostly simple wrappers around ggplot2 functions, so students don't need to grasp the intricacies of that package.
}
\author{
\strong{Maintainer}: RLadies Melbourne \email{rladiesmelbourne@gmail.com}

}
\keyword{internal}
