% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{A scatterplot}
\usage{
scatterplot(
  data,
  x,
  y,
  colour_by = NULL,
  label_x = NULL,
  label_y = NULL,
  title = NULL,
  fill_col = "black",
  line_col = "black",
  point_shape = 21
)
}
\arguments{
\item{data}{The data to plot}

\item{x}{The column to use for the x axis}

\item{y}{The column to use for the y axis}

\item{colour_by}{The column to colour the points by (optional)}

\item{label_x}{A label for the x axis}

\item{label_y}{A label for the y axis}

\item{title}{A title for the plot}

\item{fill_col}{The colour of the points}

\item{line_col}{The colour of the outline of the points}

\item{point_shape}{The shape of the points}
}
\value{
A plot
}
\description{
A scatterplot
}
\examples{
scatterplot(mpg, x = displ, y = hwy)
scatterplot(mpg, x = displ, y = hwy, colour_by = class)
scatterplot(mpg, x = displ, y = hwy, fill_col = "violet", title = "My scatterplot")
}
